+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/modular/Extensions.kt
@@ -1,0 +1,11 @@
+package org.firstinspires.ftc.teamcode.modular
+
+import com.qualcomm.robotcore.hardware.DcMotorEx
+import com.qualcomm.robotcore.hardware.DcMotorSimple
+
+fun DcMotorEx.toggleDirection() {
+    this.direction = when(this.direction) {
+        DcMotorSimple.Direction.REVERSE -> DcMotorSimple.Direction.FORWARD
+        DcMotorSimple.Direction.FORWARD -> DcMotorSimple.Direction.REVERSE
+    }
+}
